+++ test/SlowTests/Issues/RavenDB_14986.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using FastTests;
 using Orders;
+using Raven.Client.Documents;
 using Raven.Client.Documents.Indexes;
 using Tests.Infrastructure;
 using Xunit;
@@ -16,69 +17,64 @@ namespace SlowTests.Issues
         }
 
         [RavenTheory(RavenTestCategory.Indexes | RavenTestCategory.Querying)]
-        [RavenData(SearchEngineMode = RavenSearchEngineMode.Corax)]
-        public void CanSetFieldStorageNoAndFieldIndexingNoInMapReduceCorax(Options options) => CanSetFieldStorageNoAndFieldIndexingNoInMapReduce(options, simpleMapReduceErrors =>
-        {
-            Assert.Equal(25, simpleMapReduceErrors.Errors.Length);
-            Assert.True(simpleMapReduceErrors.Errors.All(x => x.Error.Contains("that is neither indexed nor stored is useless because it cannot be searched or retrieved.")));
-        });
-
-        [RavenTheory(RavenTestCategory.Indexes | RavenTestCategory.Querying)]
         [RavenData(SearchEngineMode = RavenSearchEngineMode.Lucene)]
-        public void CanSetFieldStorageNoAndFieldIndexingNoInMapReduceLucene(Options options) =>
-            CanSetFieldStorageNoAndFieldIndexingNoInMapReduce(options, simpleMapReduceErrors =>
-            {
-                Assert.Equal(25, simpleMapReduceErrors.Errors.Length);
-                Assert.True(simpleMapReduceErrors.Errors.All(x => x.Error.Contains("it doesn't make sense to have a field that is neither indexed nor stored")));
-            });
-
-        private void CanSetFieldStorageNoAndFieldIndexingNoInMapReduce(Options options, Action<IndexErrors> simpleMapAssertion)
+        public void CanSetFieldStorageNoAndFieldIndexingNoInMapReduceLucene(Options options)
         {
             using (var store = GetDocumentStore(options))
             {
-                new SimpleMapIndex().Execute(store);
-                new SimpleMapReduceIndex().Execute(store);
-
-                using (var session = store.OpenSession())
+                CanSetFieldStorageNoAndFieldIndexingNoInMapReduce(store, Indexes, simpleMapReduceErrors =>
                 {
-                    for (var i = 0; i < 25; i++)
-                        session.Store(new Company { Name = $"C_{i}", ExternalId = $"E_{i}" });
+                    Assert.Equal(25, simpleMapReduceErrors.Errors.Length);
+                    Assert.True(simpleMapReduceErrors.Errors.All(x => x.Error.Contains("it doesn't make sense to have a field that is neither indexed nor stored")));
+                });
+            }
+        }
 
-                    session.SaveChanges();
-                }
+        internal static void CanSetFieldStorageNoAndFieldIndexingNoInMapReduce(DocumentStore store, IndexesTestBase indexes, Action<IndexErrors> simpleMapAssertion)
+        {
 
-                Indexes.WaitForIndexing(store, allowErrors: true);
+            new SimpleMapIndex().Execute(store);
+            new SimpleMapReduceIndex().Execute(store);
 
-                var errors = Indexes.WaitForIndexingErrors(store);
-                Assert.Equal(2, errors.Length);
+            using (var session = store.OpenSession())
+            {
+                for (var i = 0; i < 25; i++)
+                    session.Store(new Company { Name = $"C_{i}", ExternalId = $"E_{i}" });
 
-                var simpleMapErrors = errors.Single(x => x.Name == new SimpleMapIndex().IndexName);
-                simpleMapAssertion(simpleMapErrors);
+                session.SaveChanges();
+            }
 
-                var simpleMapReduceErrors = errors.Single(x => x.Name == new SimpleMapReduceIndex().IndexName);
-                Assert.Equal(0, simpleMapReduceErrors.Errors.Length);
+            indexes.WaitForIndexing(store, allowErrors: true);
 
-                using (var session = store.OpenSession())
+            var errors = indexes.WaitForIndexingErrors(store);
+            Assert.Equal(2, errors.Length);
+
+            var simpleMapErrors = errors.Single(x => x.Name == new SimpleMapIndex().IndexName);
+            simpleMapAssertion(simpleMapErrors);
+
+            var simpleMapReduceErrors = errors.Single(x => x.Name == new SimpleMapReduceIndex().IndexName);
+            Assert.Equal(0, simpleMapReduceErrors.Errors.Length);
+
+            using (var session = store.OpenSession())
+            {
+                var results = session
+                    .Query<SimpleMapReduceIndex.Result, SimpleMapReduceIndex>()
+                    .Customize(x => x.NoCaching())
+                    .ToList();
+
+                Assert.Equal(25, results.Count);
+                Assert.True(results.All(x => x.ExternalIds != null && x.ExternalIds.Length == 1));
+
+                foreach (var result in results)
                 {
-                    var results = session
-                        .Query<SimpleMapReduceIndex.Result, SimpleMapReduceIndex>()
-                        .Customize(x => x.NoCaching())
-                        .ToList();
-
-                    Assert.Equal(25, results.Count);
-                    Assert.True(results.All(x => x.ExternalIds != null && x.ExternalIds.Length == 1));
-
-                    foreach (var result in results)
-                    {
-                        Assert.NotNull(result.ExternalIds);
-                        Assert.Equal(1, result.ExternalIds.Length);
-                        Assert.Equal(new[] { result.Name.Replace("C", "E") }, result.ExternalIds);
-                    }
+                    Assert.NotNull(result.ExternalIds);
+                    Assert.Equal(1, result.ExternalIds.Length);
+                    Assert.Equal(new[] { result.Name.Replace("C", "E") }, result.ExternalIds);
                 }
             }
         }
 
-        private class SimpleMapIndex : AbstractIndexCreationTask<Company>
+        internal class SimpleMapIndex : AbstractIndexCreationTask<Company>
         {
             public SimpleMapIndex()
             {
@@ -94,7 +90,7 @@ namespace SlowTests.Issues
             }
         }
 
-        private class SimpleMapReduceIndex : AbstractIndexCreationTask<Company, SimpleMapReduceIndex.Result>
+        internal class SimpleMapReduceIndex : AbstractIndexCreationTask<Company, SimpleMapReduceIndex.Result>
         {
             public class Result
             {
