+++ test/SlowTests/Issues/RavenDB_15568.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using FastTests;
 using Orders;
 using Raven.Client;
+using Raven.Client.Documents;
 using Raven.Client.Documents.Indexes;
 using Tests.Infrastructure;
 using Xunit;
@@ -16,7 +17,7 @@ namespace SlowTests.Issues
         {
         }
 
-        private Action<IndexErrors> _coraxAssertion = simpleMapErrors =>
+        internal static Action<IndexErrors> _coraxAssertion = simpleMapErrors =>
         {
             Assert.Equal(25, simpleMapErrors.Errors.Length);
             Assert.True(simpleMapErrors.Errors.All(x => x.Error.Contains("that is neither indexed nor stored is useless because it cannot be searched or retrieved.")));
@@ -27,11 +28,6 @@ namespace SlowTests.Issues
             Assert.Equal(25, simpleMapErrors.Errors.Length);
             Assert.True(simpleMapErrors.Errors.All(x => x.Error.Contains("it doesn't make sense to have a field that is neither indexed nor stored")));
         };
-        
-        [RavenTheory(RavenTestCategory.Indexes)]
-        [RavenData(SearchEngineMode = RavenSearchEngineMode.Corax)]
-        public void SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrorsInCorax(Options options) =>
-            SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrors<SimpleMapIndexWithDefaultFields>(options, _coraxAssertion);
 
         [RavenTheory(RavenTestCategory.Indexes)]
         [RavenData(SearchEngineMode = RavenSearchEngineMode.Lucene)]
@@ -47,29 +43,36 @@ namespace SlowTests.Issues
         [RavenData(SearchEngineMode = RavenSearchEngineMode.Lucene)]
         public void SettingDynamicFieldToNoIndexAndNoStoreShouldGenerateErrorsInLucene(Options options) =>
             SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrors<DynamicItemWhichIsNotIndexed>(options, _luceneAssertion);
-        
-        private void SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrors<TIndex>(Options options, Action<IndexErrors> assertion) where TIndex : AbstractIndexCreationTask, new()
+
+        private void SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrors<TIndex>(Options options, Action<IndexErrors> assertion)
+            where TIndex : AbstractIndexCreationTask, new()
         {
             using (var store = GetDocumentStore(options))
             {
-                new SimpleMapIndexWithDefaultFields().Execute(store);
-
-                using (var session = store.OpenSession())
-                {
-                    for (var i = 0; i < 25; i++)
-                        session.Store(new Company { Name = $"C_{i}", ExternalId = $"E_{i}" });
-
-                    session.SaveChanges();
-                }
-                WaitForUserToContinueTheTest(store);
-                Indexes.WaitForIndexing(store, allowErrors: true);
-
-                var errors = Indexes.WaitForIndexingErrors(store);
-                Assert.Equal(1, errors.Length);
-
-                var simpleMapErrors = errors.Single(x => x.Name == new SimpleMapIndexWithDefaultFields().IndexName);
-                assertion(simpleMapErrors);
+                SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrorsInternal<TIndex>(store, Indexes, assertion);
             }
+        }
+
+        internal static void SettingDefaultFieldsToNoIndexAndNoStoreShouldGenerateErrorsInternal<TIndex>(DocumentStore store, IndexesTestBase indexes, Action<IndexErrors> assertion) where TIndex : AbstractIndexCreationTask, new()
+        {
+            new SimpleMapIndexWithDefaultFields().Execute(store);
+
+            using (var session = store.OpenSession())
+            {
+                for (var i = 0; i < 25; i++)
+                    session.Store(new Company { Name = $"C_{i}", ExternalId = $"E_{i}" });
+
+                session.SaveChanges();
+            }
+            WaitForUserToContinueTheTest(store);
+            indexes.WaitForIndexing(store, allowErrors: true);
+
+            var errors = indexes.WaitForIndexingErrors(store);
+            Assert.Equal(1, errors.Length);
+
+            var simpleMapErrors = errors.Single(x => x.Name == new SimpleMapIndexWithDefaultFields().IndexName);
+            assertion(simpleMapErrors);
+
         }
 
         private class DynamicItemWhichIsNotIndexed : AbstractIndexCreationTask<Company>
@@ -85,7 +88,7 @@ namespace SlowTests.Issues
         
         //A field `Name` that is neither indexed nor stored is useless because it cannot be searched or retrieved.
 
-        private class SimpleMapIndexWithDefaultFields : AbstractIndexCreationTask<Company>
+        internal class SimpleMapIndexWithDefaultFields : AbstractIndexCreationTask<Company>
         {
             public SimpleMapIndexWithDefaultFields()
             {
