+++ test/SlowTests/Server/Documents/PeriodicBackup/PeriodicBackupSlowTests.cs
@@ -2939,75 +2939,6 @@ namespace SlowTests.Server.Documents.PeriodicBackup
             }
         }
 
-        // Performing backup Delay to the time:
-        [InlineData(1)] // until the next scheduled backup time.
-        [InlineData(5)] // after the next scheduled backup.
-        [NightlyBuildTheory, Trait("Category", "Smuggler")]
-        public async Task ShouldProperlyPlaceOriginalBackupTimePropertyWithDelay(int delayDurationInMinutes)
-        {
-            const string fullBackupFrequency = "*/2 * * * *";
-            var backupPath = NewDataPath(suffix: "BackupFolder");
-
-            using (var server = GetNewServer())
-            using (var store = GetDocumentStore(new Options { Server = server }))
-            {
-                using (var session = store.OpenAsyncSession())
-                    await Backup.FillDatabaseWithRandomDataAsync(databaseSizeInMb: 1, session);
-
-                var database = await server.ServerStore.DatabasesLandlord.TryGetOrCreateResourceStore(store.Database);
-                Assert.NotNull(database);
-                await Backup.HoldBackupExecutionIfNeededAndInvoke(database.PeriodicBackupRunner.ForTestingPurposesOnly(), async () =>
-                {
-                    WaitForValue(() =>
-                    {
-                        var now = DateTime.Now;
-                        return now.Minute % 2 == 0 && now.Second <= 10;
-                    },
-                       expectedVal: true,
-                       timeout: (int)TimeSpan.FromMinutes(2).TotalMilliseconds,
-                       interval: (int)TimeSpan.FromSeconds(1).TotalMilliseconds
-                   );
-
-                    var config = Backup.CreateBackupConfiguration(backupPath, fullBackupFrequency: fullBackupFrequency);
-                    var taskId = await Backup.UpdateConfigAndRunBackupAsync(server, config, store, opStatus: OperationStatus.InProgress);
-                    // Let's delay the backup task
-                    var taskBackupInfo = await store.Maintenance.SendAsync(new GetOngoingTaskInfoOperation(taskId, OngoingTaskType.Backup)) as OngoingTaskBackup;
-                    Assert.NotNull(taskBackupInfo);
-                    Assert.NotNull(taskBackupInfo.OnGoingBackup);
-                    Assert.NotNull(taskBackupInfo.OnGoingBackup.StartTime);
-
-                    var delayDuration = TimeSpan.FromMinutes(delayDurationInMinutes);
-                    var delayUntil = DateTime.Now + delayDuration;
-                    await store.Maintenance.SendAsync(new DelayBackupOperation(taskBackupInfo.OnGoingBackup.RunningBackupTaskId, delayDuration));
-
-                    // There should be no OnGoingBackup operation in the OngoingTaskBackup
-                    await WaitForValueAsync(async () =>
-                    {
-                        var afterDelayTaskBackupInfo = await store.Maintenance.SendAsync(new GetOngoingTaskInfoOperation(taskId, OngoingTaskType.Backup)) as OngoingTaskBackup;
-                        return afterDelayTaskBackupInfo is { OnGoingBackup: null };
-                    }, true);
-
-                    var backupStatus = (await store.Maintenance.SendAsync(new GetPeriodicBackupStatusOperation(taskId))).Status;
-                    Assert.NotNull(backupStatus);
-                    Assert.NotNull(backupStatus.DelayUntil);
-                    Assert.NotNull(backupStatus.OriginalBackupTime);
-
-                    var nextFullBackup = GetNextBackupOccurrence(new NextBackupOccurrenceParameters
-                    {
-                        BackupFrequency = fullBackupFrequency,
-                        Configuration = config,
-                        LastBackupUtc = taskBackupInfo.OnGoingBackup.StartTime.Value
-                    });
-                    Assert.NotNull(nextFullBackup);
-
-                    Assert.Equal(backupStatus.OriginalBackupTime,
-                        delayUntil < nextFullBackup
-                            ? taskBackupInfo.OnGoingBackup.StartTime    // until the next scheduled backup time.
-                            : nextFullBackup.Value.ToUniversalTime());  // after the next scheduled backup.
-                }, tcs: new TaskCompletionSource<object>(TaskCreationOptions.RunContinuationsAsynchronously));
-            }
-        }
-
         [Fact, Trait("Category", "Smuggler")]
         public async Task ShouldHaveFailoverForFirstBackupInNewBackupTask()
         {
