+++ test/SlowTests/Voron/Issues/RavenDB_19278.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using FastTests.Voron;
 using Sparrow.Platform;
 using Sparrow.Utils;
+using Tests.Infrastructure;
 using Voron;
 using Voron.Global;
 using Voron.Impl.Journal;
@@ -146,8 +147,13 @@ public class RavenDB_19278 : StorageTest
         StartDatabase();
     }
 
+    [NightlyBuildFact]
+    public void PartialRecoveryMustUpdateEnvironmentHeaderAndEraseCorruptedDataInJournal_Nightly()
+    {
+        PartialRecoveryMustUpdateEnvironmentHeaderAndEraseCorruptedDataInJournal(true, false, false);
+    }
+
     [Theory]
-    [InlineData(true, false, false)]
     [InlineData(false, true, false)]
     [InlineData(false, false, false)]
     [InlineData(true, true, false)]
